Some cleanup of report generation
Move table creation to Report::Table and subclass from there.
