+++ lib/brakeman/report/report_tabs.rb
@@ -1,6 +1,8 @@
+require 'brakeman/report/report_table'
+
 #Generated tab-separated output suitable for the Jenkins Brakeman Plugin:
 #https://github.com/presidentbeef/brakeman-jenkins-plugin
-class Brakeman::Report::Tabs < Brakeman::Report::Base
+class Brakeman::Report::Tabs < Brakeman::Report::Table
   def generate_report
     [[:generic_warnings, "General"], [:controller_warnings, "Controller"],
       [:model_warnings, "Model"], [:template_warnings, "Template"]].map do |meth, category|
