+++ lib/brakeman/warning.rb
@@ -213,7 +213,7 @@ class Brakeman::Warning
 
   #Generates a hash suitable for inserting into a table
   def to_row type = :warning
-    @row = { "Confidence" => self.confidence,
+    @row = { "Confidence" => TEXT_CONFIDENCE[self.confidence],
       "Warning Type" => self.warning_type.to_s,
       "Message" => self.message }
